# Optimize GPU performance of slice and concat layers