This PR makes two changes to the slice and concatenation layers:

- Use Hydrogen `simple_buffer`s for GPU workspaces rather than Thrust `device_vector`s. Thrust synchronizes the CUDA stream whenever it creates a `device_vector`, hurting kernel pipelining.
- CUDA kernels use nested loops to iterate through tensors. Previously we had a single loop and computed tensor positions with divides and moduli, but the kernel runtime was dominated by integer arithmetic. The new approach is optimized for the slice layers in multi-head attention, although it may be susceptible to bad performance with certain tensor shapes (if there are many children/parents, if an NCHW tensor is dominated by the N dimension, if the children/parent tensors have a wide range of sizes).

Timings with a toy model with repeated slices and concats (1 proc on Pascal):

| Commit | Mini-batch time |
|----|----|
| 9a9e31c (develop) | 0.0284 s |
| 14e00c4 | 0.0281 s |
| df70e84 | 0.0128 s |

Timings with Transformer (1 Pascal node):

| Commit | Mini-batch time |
|----|----|
| 9a9e31c (develop) | 0.276 s |
| 14e00c4 | 0.286 s |
| df70e84 | 0.259 s |

[Bamboo](https://lc.llnl.gov/bamboo/browse/LBANN-TIM284-1) doesn't show any new failures.

Notes for future optimizations:
- We map the CUDA kernel grid to the C, H, and W dimensions of an NCHW tensor. Is it worthwhile coming up with a separate implementation if N is large?
- We don't get full occupancy on the SMs because the kernel requires too many registers per thread (48 on Pascal). I tried staging some of the values in shared memory, but that ended up increasing the number of required registers.
- The kernel runtime is still dominated by integer arithmetic.
- In a toy model with repeated slices and concats, kernel pipelining is perfect.