@timmoon10 2020-06-01
Timings with the ATOM RNN model:
| Commit | Mini-batch time |
| --- | --- |
| 7e3d1bf (develop) | 0.0826 s |
| df70e84 | 0.0758 s |