@timmoon10 2020-06-10
My original toy model (repeated slices and concats) doesn't scale, likely because it is small enough for the runtime to be dominated by communication. I bumped up the number of layers and the mini-batch size and ran on Pascal:

| Nodes | Mini-batch time |
|---|---|
|1 | 0.260 sec |
| 2 | 0.166 sec |
| 4 | 0.134 sec |
| 8 | 0.110 sec |